2021/8/15 对feign接口需要和不需要自定义fallback进行分离

+++ cloud-yblog-commons/src/main/java/com/boot/utils/Commons.java
@@ -101,7 +101,7 @@ public class Commons {
      * @param aid
      * @return
      */
-    public static String permalink(Integer aid) {
+    public static String permalink(Long aid) {
         return site_url("/article/" + aid.toString());
     }
